requirements.txt
customtkinter==5.2.2
pillow==10.4.0
pandas==2.2.2
openpyxl==3.1.5